Add build scope to prevent auto rebuild loop
If this option is not set, protoc builder will be triggered if "Build
automatically" is set no matter what resource is updated.  However, the
output for protoc, /gen/com/mwr/jdiesel/Protobuf.java, is a resource for
java builder which will trigger rebuild.  In this case an infinite build
loop happens.  The solution is set the build scope for protoc builder to
common/protobuf.proto only.